// src/model/product.class.ts
export default class ProductDto {
  $key: string;
  id: number;
  type: string;
  brand: string;
  name: string;
  description: string;
  picture: string;
  handle: string;
  steel: string;
  size: string;
  youtube_ref: string;
  price: number;
  comment: string;
  promo: boolean;
  old_price: number;
  in_stock: boolean;
  type_brand: string;
}